refactor: UI 수정 및 svg 이미지 동적으로 사용하도록 수정
